Move TransactionGraphSearch into trader demo
